feat: cr and tdl to separate resolver files
